Break out cli into lazily loaded sub modules
